fix atan2
